Add support for setting custom bitmap cursors
Right now you can set the mouse cursor shape in a TouchArea to a
pre-defined list of cursor shapes, but sometimes you have your own
custom cursor set for actions that aren't in the CSS spec.

This change introduces new API to set a mouse cursor to an Image, using
the same property (mouse-cursor) and a changeable hotspot.